Update samples/java/play-authenticate-usage/conf/play-authenticate/mine.conf
Little omission. It would go un-noticed unless you try to compile from source, play-authenticate code together with play-authenticate-usage. If you do you'd get an error. 
Not sure from where it is loaded if using the play-authenticated as a module, and why it does not give any errors in that case.